Network Config Update + Mouse Look update to be done via IBeforeUpdate.BeforeUpdate()
Using IBeforeUpdate allows the input to be handled before every Unity Update. This is important as we want to have our inputs be smooth, and not update over the FixedNetworkUpdate() which will introduce some input delay due to the FixedNetworkUpdate not being tied to frame updates.

- Also in the NetworkProjectConfig, we set the Prediction down to 2 ticks, as well as Hardsnap down to 100% for better clientside responsiveness. (more prediction means more smoothing in player movements / rotations) and we want it to be snappy and responsive while being interpolated.

So now the mouse inputs are not tied to the server tick rate. (Before we had to set it high enough at around 120 to get a decent responsiveness, but that was not the proper way to achieve so. It should be done via IBeforeUpdate as noted above)